refactor manual impl of sep double exp model in shared crate
